Change the way that X-Ray exporter annotation converter work (#31732)
Description:
In the past, X-Ray doesn’t support “.”. So we have a translator in x-ray
export to translates it to “_” before sending traces to X-Ray Service.

To match otel naming style, x-ray service team decide to change their
service to support both "." type and "" type of naming. In this case the
translator that translate "." to "" is no-longer needed. This PR change
the way this translator work

X-Ray PMs agree on rolling out this change by using feature-gate

Testing:
Unit test