[chore][receiver/sqlserver] Add direct connection config options (#32176)
**Description:** <Describe what has changed.>
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
This adds config options for direct connection to a SQL server instance.
For the sake of PR size, this adds no functionality, it's simply config
options and validation.

**Link to tracking Issue:** <Issue number if applicable>
This was originally part of
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/31915,
but I'm breaking this out to make the original PR a more manageable
size.


https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/30297

**Testing:** <Describe what testing was performed and which tests were
added.>
Existing tests and added tests are all passing.

**Documentation:** <Describe the documentation added.>
No documentation on purpose. The added config options currently have no
impact on functionality, so users shouldn't be aware of this, and
shouldn't try to use them yet.