[internal/metadataproviders/azure] Fix memory leak on AKS (#32574)
**Description:** <Describe what has changed.>
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
From [`client.Do` documentation](https://pkg.go.dev/net/http#Client.Do),
a client's response can safely be ignored when an error is returned, but
must be handled in all other cases, even when a response code is not
2XX. This fixes the internal AKS detection to properly close the
response body in the case of a non-200 status returned.

This also enables `goleak` on the `internal/metadata/azure` package
which helps ensure no goroutines are being leaked. `goleak` is what
detected this bug.

**Link to tracking Issue:** <Issue number if applicable>
#30438

**Testing:** <Describe what testing was performed and which tests were
added.>
All existing tests are passing as well as added `goleak` check.