[chore] [pkg/stanza] Pass telemetry settings by value (#32672)
**Description:** <Describe what has changed.>
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
Telemetry settings in general are passed by value in the project. This
commit aligns the stanza Builder interface with this convention. This
will also ensures that Logger can be overwritten without changing the
original object.
Follow up of
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/32662.
cc-ing @djaglowski who pointed that out.

**Link to tracking Issue:** <Issue number if applicable>

**Testing:** <Describe what testing was performed and which tests were
added.> Same as
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/32662

**Documentation:** <Describe the documentation added.>

Signed-off-by: ChrsMark <chrismarkou92@gmail.com>