[receiver/vcenter] Adds Suspended VM Counts to `vcenter.cluster.vm.count` (#32804)
**Description:** <Describe what has changed.>
Currently the metric `vcenter.cluster.vm.count` does not include a count
for suspended VMs. Instead it lumps together this count with the value
it reports for "poweredOn" VMs.

Here I'm modifying the relevent metric attribute to allow for a
"suspended" value, and updating the metric to report this count.

I also added a minor capitalization fix for the metadata description.

**Link to tracking Issue:** <Issue number if applicable>
#32803 

**Testing:** <Describe what testing was performed and which tests were
added.>
Unit/integration tests updated and tested. Local environment tested.

**Documentation:** <Describe the documentation added.>
New documentation generated based on the metadata.