[jsonlogencodingextension] Add Modes for Body with Attributes (#32722)
**Description:** <!--Ex. Fixing a bug - Describe the bug and how this
fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->

The goal of this PR is to add configuration to the JSON Encoding method
to have metadata from resources and attributes in logs and output those
into a json attribute. This also will output the body of the log in the
`body` attribute. This follows a similar pattern that the AWS Cloudwatch
Log Exporter uses but removes any of the AWS specific logic and makes
this a more generalized solution

**Link to tracking Issue:** <Issue number if applicable>
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/32679

**Testing:** <Describe what testing was performed and which tests were
added.> Unit Tests were added to validate that the output is desired and
that the existing logic remains the same.

**Documentation:** <Describe the documentation added.> Documentation
regarding the usage of the configuration as well as the output to expect
is added.

---------

Signed-off-by: Blake R <85771645+blakeromano@users.noreply.github.com>