[receiver/vcenter] Fixes Client Virtual App Call Error When No Virtual Apps (#33102)
**Description:** <Describe what has changed.>
Virtual Apps might not exist in an environment. Currently a `NotFound`
error for the internal client is thrown when this condition occurs. I
modified this to simply return an empty array, because this should
simply mean that it did not find any Virtual Apps. I also added more
client unit tests to test for these types of conditions.

**Link to tracking Issue:** <Issue number if applicable>
#33073 (loosely tied to this. At the very least this issue made me
discover this problem)

**Testing:** <Describe what testing was performed and which tests were
added.>
Added new unit tests which showed the error. Fix caused the new unit
test to pass.

**Documentation:** <Describe the documentation added.>
N/A