[cmd/telemetrygen] Add exemplars support for telemetrygen (#33320)
This PR adds support for passing a traceID and spanID to be used as
exemplars on the generated data points. This was verified with an e2e
test using OTel Collector and the debug exporter in detailed verbosity,
resulting in this:

```
2024-05-31T11:41:51.378+0200    info    ResourceMetrics #0
Resource SchemaURL: https://opentelemetry.io/schemas/1.13.0
ScopeMetrics #0
ScopeMetrics SchemaURL: 
InstrumentationScope  
Metric #0
Descriptor:
     -> Name: gen
     -> Description: 
     -> Unit: 
     -> DataType: Gauge
NumberDataPoints #0
StartTimestamp: 1970-01-01 00:00:00 +0000 UTC
Timestamp: 2024-05-31 09:41:51.374564535 +0000 UTC
Value: 0
Exemplars:
Exemplar #0
     -> Trace ID: ae87dadd90e9935a4bc9660628efd569
     -> Span ID: 5828fa4960140870
     -> Timestamp: 2024-05-31 09:41:51.374432029 +0000 UTC
     -> Value: 1
```

Given that I needed to do the validation of the traceID/spanID on the
metrics config as well, I refactored the current validation done in the
logging part of the command so that it's now part of the common package.

Signed-off-by: Juraci Paixão Kröhling <juraci@kroehling.de>

---------

Signed-off-by: Juraci Paixão Kröhling <juraci@kroehling.de>