[chore] Fix .github/CODEOWNERS (#33677)
`main` is currently failing the [`checks`
action](https://github.com/open-telemetry/opentelemetry-collector-contrib/actions/runs/9588994799/job/26442043847),
at the `Gen CODEOWNERS` step. This applies the required change.

This was broken by
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/33514,
which added a `metadata.yaml` file for the `pkg/stanza/fileconsumer`
package, but didn't add an entry to the CODEOWNERS file.