[processor/resourcedetection] Fetch CPU info only if required (#33774)
**Description:** <Describe what has changed.>
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
This PR changes the system resource detector so as to only try to fetch
the CPU info when required. The CPU info attributes are disabled by
default so we should only fetch this information when at least one of
those is enabled.

**Link to tracking Issue:** <Issue number if applicable>
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/33768

**Testing:** <Describe what testing was performed and which tests were
added.> Added unti-tests

**Documentation:** <Describe the documentation added.> ~

/cc @mx-psi

---------

Signed-off-by: ChrsMark <chrismarkou92@gmail.com>