[chore] Source Timeout for Providers (#33958)
**Description:** <Describe what has changed.>

Datadog exporter tests have an unnecessary 30 second wait while it waits
for various metadata service calls to timeout (this is the default TCP
timeout in the go network stack). Also had to modify the ec2 provider so
that it respected the context. Most of the changes are just propagating
this new timeout to the provider, and propagating this context around. I
made the decision to differentiate the timeout context and the overall
context for the chainProvider - these could be brought together but it
looks like we had existing logic for a context timeout vs a
chainProvider context cancellation so I wanted to respect that split.

**Testing:** 
mostly stuck to a 31 second timeout on existing tests to not change
behavior. Added a tighter timeout for some of the slower tests.
Added some tests for the timeout logic

---------

Co-authored-by: Pablo Baeyens <pbaeyens31+github@gmail.com>