[otelarrowreceiver] Avoid one unnecessary span error at stream EOF (#34175)
**Description:** We have been using the OTel-Arrow components in
production and begun to monitor the component's health using span data.
There are spans coming from otelgrpc instrumentation and there are
explicit spans in both components.

This adds a new test that exercises multiple streams between the
exporter and receiver components. At the end of the test, it checks for
no span errors, with one exception discussed in
https://github.com/open-telemetry/opentelemetry-go-contrib/issues/2644.
We will not modify the logic of these components to avoid stream
cancelation, we will lobby for stream cancelation not to register as
span errors.

The receiver is modified to avoid setting Span errors for EOF and
Cancelation following its existing logic for setting log severity for
the same condition. As a result, the `otel_arrow_stream_inflight` span
will not show errors for the final `Recv()` as a stream shuts down.

**Link to tracking Issue:** Part of
https://github.com/open-telemetry/otel-arrow/issues/227.

**Testing:** A new test is added.