[connector/otlpjson] Add connector's implementations (#34249)
**Description:** <Describe what has changed.>
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
This is the 2nd PR for the new `otlpjson` connector
(https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/34208).
Adds the `ConsumeLogs` methods for the logs, metrics and traces
respective connector types.

At the moment, the connector retrieves the otlp json from the incoming
Log's `Body`. This is not configurable yet. It can be decided if that
needs to be.

In addition, there is no attribute/metadata extraction/transfer from the
incoming Log. This can be added in future iterations.

**Link to tracking Issue:** <Issue number if applicable>
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/34208
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/34239

**Testing:** <Describe what testing was performed and which tests were
added.> Added.

**Documentation:** <Describe the documentation added.> ~

Signed-off-by: ChrsMark <chrismarkou92@gmail.com>