[processor/k8sattributes] Add `container.image.repo_digests` (#34031)
**Description:**

Adds support for extracting canonical repo digest image references for a
container from k8s attributes

**Link to tracking Issue:** #34029 

**Testing:**

1. Adds unit tests for extracting the k8s `.imageID` field from
container status API if it is of the `reference.Canonical` form, and
storing it in a `ImageRepoDigest` string field. (Note: k8s will only
ever return one value here, although the semantic conventions for
`container.image.repo_digests` define it to be a `Slice`. We convert
into a Slice lazily when we populate the pcommon data structure.
2. Adds unit tests for including `container.image.repo_digests` in rules
for extraction.
3. Includes `container.image.repo_digests` in the e2e tests for
k8sattributes processor.


**Documentation:**

Added `container.image.repo_digests` as an example of attributes that
can be extracted from a container

---------

Co-authored-by: Christos Markou <chrismarkou92@gmail.com>