[exporter/clickhouse] Add `compress` option to config, enabled by default (#34365)
**Description:**

This change adds a new `compress` option to the config and sets it to
`lz4` by default.

In the current version of the exporter, users must know to provide
`compress` in the DSN URL to gain the network performance benefits of
compression. The only way they would have known this before is if they
copied the sample from the README, but this is likely replaced when they
paste their server address.

ClickHouse has excellent compression for storage and network. It is
recommended to enable it for clients such as the OTel exporter to
improve performance.

In summary:
- Added `compress` field to config
- `endpoint` (DSN URL) and `connection_params` takes priority
- If left empty from all sources, will default to `lz4`
- Valid options are based on the underlying `clickhouse-go` driver:
`none` (disabled), `zstd`, `lz4` (default), `gzip`, `deflate`, `br`,
`true` (lz4).

The `true` option comes from an older version of `clickhouse-go` and is
an alias for `lz4`. To prevent unexpected changes in behavior, I have
manually re-added this check to the config parser instead of assuming
the driver will still interpret it as `lz4`.

**Testing:**
- Updated unit tests for DSN + config parsing
- Ran integration tests locally

**Documentation:**
- Updated README config options list + sample
- Added changelog

---------

Co-authored-by: Pablo Baeyens <pbaeyens31+github@gmail.com>