[exporter/signalfx] Fix goroutine leaks (#32781)
**Description:** 
This change is a refactor to allow us to enable `goleak` checks in each
of the exporter's sub-directories, as well as the signalfx receiver
(which uses the signalfxexporter in a test). The main idea is to use
`start`/`shutdown` for each package's functionality, rather than relying
on the top level context cancel of the exporter. The addition of start
and shutdowns for each package means we can more closely control the
lifecycle of each as needed, and enable `goleak` for each package.

The memory leaks being fixed are:
1. Call the `TTLMap.Shutdown` so the ticker is properly stopped:
`t.prevPts.Shutdown()`.
2. Don't block waiting for a request if the context is cancelled.

**Link to tracking issue:**
#30438

---------

Co-authored-by: Sean Marciniak <30928402+MovieStoreGuy@users.noreply.github.com>