[extension/healthcheckv2] Update readme to reflect current status (#34913)
There was a warning added to the readme indicating that the health check
v2 extension was not ready for use. This was added because the extension
was sliced up into a number of smaller PRs that were incrementally
reviewed and merged. That process finished a while ago, but was followed
by a component status refactor on core. Now that that work is complete,
the extension is available for experimentation purposes for those who
want to build a custom collector binary. This PR updates the readme to
reflect the current status of the extension.