[chore] Fix "Device or resource busy" error on Windows CI (#35192)
**Description:**
The error "Device or resource busy" happens intermittently when running
the `build-and-test-windows` workflow on CI. Examining a few instances
of the error it becomes clear that the concurrent run of component tests
trigger concurrent builds of `gotestsum`. This can cause one of the runs
to clash with the other and fail to build the `gotestsum` target.

The fix is to add a explicit step to build `gotestsum` in the CI
workflow. This way the local dev modus operandi is not altered (they can
still rely on `make test` to build any tool necessary for the test). If
any other tools are added to the `test` target they should also be added
to the workflow on Windows.

**Link to tracking Issue:**
Fix #34691

**Testing:**
Validated on repository fork.

**Documentation:**
N/A