[pkg/stanza] Windows input operator should resubscribe when remote host restarts (#35175)
**Description:** <Describe what has changed.>
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->

While collecting Windows Event Logs from a remote host, if that host
restarts or shuts off, the collector begins logging errors and stops
collecting WE logs, even after the remote host is back up. Restarting
the collector fixes this.

To fix this without restarting the collector, the stanza input operator
needs to resubscribe if the remote host restarts.

Error log:
```
"Failed to read events from subscription","error":"The handle is invalid."
```


**Link to tracking Issue:** <Issue number if applicable>

**Testing:** <Describe what testing was performed and which tests were
added.>

Manually tested with a build of the collector, and confirmed logs
continue being collected after the remote host restarts.

**Documentation:** <Describe the documentation added.>

N/A, this change is expected behavior