[chore][Testbed] - Try flushing the data before exiting in load generator (#35211)
**Description:** In the `LoadGenerator`, if we encounter a non-permanent
error and are in the process of retrying, we should not immediately exit
upon receiving a stop signal. Instead, we need to first flush the
existing data and then proceed to exit the LoadGenerator. This is
necessary for stress test cases and we need to validate the data
received and data sent.

---------

Co-authored-by: Daniel Jaglowski <jaglows3@gmail.com>