[receiver/sqlserver] Add computer name resource attribute to relevant metrics (#35040)
**Description:** <Describe what has changed.>
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
The computer name resource attribute is relevant for all metrics coming
from the SQL server receiver when directly connected to the DB. This
change records the resource attribute with all remaining metrics that it
wasn't recorded with yet.

**Link to tracking Issue:** <Issue number if applicable>
This was discovered at the same time as
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/35036,
but I split out the changes to hopefully make them easier to review.

**Testing:** <Describe what testing was performed and which tests were
added.>
Tests have been updated and are passing.