[processor/geoip] skip error if geo metadata is not found (#35278)
**Description:** <Describe what has changed.>
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.--> 
If a provider does not found any associated metadata to the given IP,
the processor will continue the processing instead of returning the
error. Nonetheless, the error will be logged when debug telemetry level
is enabled.

**Link to tracking Issue:** <Issue number if applicable>
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/35047

**Testing:** <Describe what testing was performed and which tests were
added.> Add a testdata case for IP `1.2.3.5` which is not available in
any of the providers (maxmind neither mocked provider)

**Documentation:** <Describe the documentation added.>